requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.6
llama_cpp_python==0.2.20
numpy==1.26.2
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
sse-starlette==1.8.2
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
